# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Error Database API
  description: |
    A comprehensive API for the Error Database application. This API provides endpoints for managing error codes, solutions, user authentication, and advanced features like bookmarks, notifications, achievements, webhooks, and data export.

    ## Features
    - **Error Management**: Create, read, update, and search error codes
    - **Solution Management**: Submit, vote on, and verify solutions
    - **User Management**: Authentication, profiles, and achievements
    - **Bookmarks**: Save solutions for later reference
    - **Notifications**: Real-time in-app notifications
    - **Webhooks**: External integrations with event subscriptions
    - **Analytics**: Search analytics and usage statistics
    - **Export**: Data export in JSON and CSV formats

    ## Authentication
    All protected endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
  version: 1.0.0
  contact:
    name: Error Database Team
    email: support@errordb.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3010/api
    description: Local development server
  - url: https://api.errordb.com/api
    description: Production server
...
paths:
...
  /search:
    get:
      tags:
        - Search
      summary: Advanced search with filters
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - name: applicationId
          in: query
          schema:
            type: string
            format: uuid
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: hasSolutions
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [relevance, newest, popular, solutions]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      errors:
                        type: array
                        items:
                          $ref: '#/components/schemas/ErrorCode'
                      totalCount:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
                      totalPages:
                        type: integer
  /search/suggestions:
    get:
      tags:
        - Search
      summary: Get search suggestions/autocomplete
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 5
      responses:
        '200':
          description: Search suggestions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          enum: [error, popular, application, category]
                        value:
                          type: string
                        label:
                          type: string
  /search/history:
    get:
      tags:
        - Search
      summary: Get user search history
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Search history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SearchHistoryItem'

    delete:
      tags:
        - Search
      summary: Clear all search history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: History cleared
  /search/history/{id}:
    delete:
      tags:
        - Search
      summary: Delete a specific search history item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Item deleted
  /search/trends:
    get:
      tags:
        - Search
      summary: Get search trends (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 7
      responses:
        '200':
          description: Search trends
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        query:
                          type: string
                        count:
                          type: integer
        '403':
          description: Admin access required

  # Bookmarks
...
  /webhooks:
    get:
      tags:
        - Webhooks
      summary: Get user webhooks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User webhooks
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'

    post:
      tags:
        - Webhooks
      summary: Create a webhook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - events
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                    enum:
                      - error.created
                      - error.updated
                      - solution.created
                      - solution.verified
                      - solution.upvoted
                      - user.registered
                      - category.created
                      - application.created
      responses:
        '201':
          description: Webhook created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      webhook:
                        $ref: '#/components/schemas/Webhook'
                      secret:
                        type: string
  /webhooks/{id}:
    patch:
      tags:
        - Webhooks
      summary: Update a webhook
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                isActive:
                  type: boolean
      responses:
        '200':
          description: Webhook updated

    delete:
      tags:
        - Webhooks
      summary: Delete a webhook
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Webhook deleted
  /webhooks/{id}/regenerate-secret:
    post:
      tags:
        - Webhooks
      summary: Regenerate webhook secret
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Secret regenerated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      secret:
                        type: string
...
  /webhooks/events/types:
    get:
      tags:
        - Webhooks
      summary: Get available webhook event types
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Event types
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        key:
                          type: string
                        value:
                          type: string
                        description:
                          type: string

  # Export
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorCode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        viewCount:
          type: integer
        application:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            category:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
        solutions:
          type: array
          items:
            $ref: '#/components/schemas/Solution'
        _count:
          type: object
          properties:
            solutions:
              type: integer
    Solution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        solutionText:
          type: string
        score:
          type: integer
        upvotes:
          type: integer
        downvotes:
          type: integer
        isVerified:
          type: boolean
        author:
          type: object
          properties:
            id:
              type: string
            username:
              type: string
            displayName:
              type: string
        createdAt:
          type: string
          format: date-time
...
    Webhook:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        lastDeliveryAt:
          type: string
          format: date-time
        lastDeliveryStatus:
          type: string
        failureCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SearchHistoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        query:
          type: string
        filters:
          type: object
        resultCount:
          type: integer
        clickedResultId:
          type: string
        createdAt:
          type: string
          format: date-time
...